requirements.txt
Flask==3.0.3
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
Pillow==10.2.0
mediapipe
opencv-python==4.10.0.84
pydub==0.25.1
librosa==0.10.2.post1